requirements.txt
google-adk>=1.18.0
litellm>=1.81.0
chromadb>=1.5.0
pydantic>=2.0
requests>=2.32.0
langdetect>=1.0.9
python-dotenv>=1.0.0
json-repair>=0.30.0
langwatch>=0.2.10
